requirements.txt
flask
loguru
yt_dlp
yt-dlp[default,curl-cffi]
waitress
python-magic
tensorflow
scikit-learn
numpy
Pillow